Collect thesis by the group members in a page
The thesis entries are shown on the author and advisor pages. For
advisors, only the latest 5 theses are shown.
Also show a list on the front-page of the latest.

@birocoles I couldn't find an abstract for your PhD thesis. I also
changed the links in your page to point to the thesis page on the
website instead of the pdf on the ON site.

I also couldn't find the thesis of Fillipe Claudio. The link on the ON
site is broken. If you have a copy, could you add it and paste the
abstract on the thesis page `thesis/siqueira-phd.md`?