Phd theses (#37)
* replace the pdf file of my phd thesis and correct small error in text

* replace the pdf of the melo's thesis and include the link to Sucupira platform

* replace the pdf file of carlos' thesis, include the Sucupira link, include the generated paper

* include the pdf file and abstract of the siqueira's thesis

* include the Sucupira link

* include the Sucupira link and the related paper

* include the sucupira link and the abstract in Andre's thesis

* Add metadata entry for a link to Sucupira

You only need to specify the ID number at the end of the Sucupira entry
link for a thesis. The template will add a link to the Info side bar.

* Add related_papers and related_thesis metadata

These entries are used to link together papers and thesis that are
related.
I have put the links to the papers and thesis that should be linked.